requirements.txt
beautifulsoup4==4.12.3
PyYAML==6.0.1
requests==2.31.0
Flask==2.3.3
gunicorn==21.2.0
